util_saml: correctly account for gss lengths
The version of XMLString::transcode that takes a length is an output
parameter length not an input length. It was reading past end of
buffer on the input and was producing garbage output. That version is
deprecated; use the version that allocates.  Copy the GSS attribute
names and values so we can properly null terminate.