Disable Shibboleth other than resolver
Debian is moving from apache 2.2 to Apache 2.4. Currently this breaks
the Shibboleth SP packaging.  That's being fixed within Debian, but by
the time that is fixed they will be building a version of Shibboleth
that is useful to us there.  Unless we need to make additional changes, don't build Shibboleth for now.